requirements.txt
apache-airflow-providers-postgres
pandas
psycopg2-binary
dbt-core==1.7.0
dbt-postgres==1.7.0
